Extract AssertRequest into a reusable trait
