+++ tests/Insightly/Resources/InsightlyContactResourceTest.php
@@ -12,11 +12,13 @@ use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\TestCase;
 use Ramsey\Uuid\Uuid;
+use Tests\AssertRequest;
 use Tests\MockCrmClient;
 
 final class InsightlyContactResourceTest extends TestCase
 {
     use MockCrmClient;
+    use AssertRequest;
 
     private InsightlyContactResource $resource;
 
@@ -72,14 +74,5 @@ final class InsightlyContactResourceTest extends TestCase
             ->with(self::callback(fn ($actualRequest): bool => self::assertRequestIsTheSame($expectedRequest, $actualRequest)));
 
         $this->resource->delete(42);
-    }
-
-    private static function assertRequestIsTheSame(Request $expected, Request $actual): bool
-    {
-        self::assertEquals($expected->getHeaders(), $actual->getHeaders());
-        self::assertEquals($expected->getMethod(), $actual->getMethod());
-        self::assertEquals($expected->getBody()->getContents(), $actual->getBody()->getContents());
-
-        return true;
     }
 }
